block: annotate refault stalls from IO submission
psi tracks the time tasks wait for refaulting pages to become
uptodate, but it does not track the time spent submitting the IO. The
submission part can be significant if backing storage is contended or
when cgroup throttling (io.latency) is in effect - a lot of time is
spent in submit_bio(). In that case, we underreport memory pressure.

Annotate submit_bio() to account submission time as memory stall when
the bio is reading userspace workingset pages.

Change-Id: I2c1b353cf1d57979531c4462797375e53512d30b
Tested-by: Suren Baghdasaryan <surenb@google.com>
Signed-off-by: Johannes Weiner <hannes@cmpxchg.org>
Signed-off-by: Jens Axboe <axboe@kernel.dk>
(1. Manually resolved BIO_WORKINGSET being definition instead of enum.
 2. Add bio_set_flag call into bio_add_page instead of not yet existing
Signed-off-by: Suren Baghdasaryan <surenb@google.com>
Signed-off-by: mydongistiny <jaysonedson@gmail.com>